feat(18458): Move MemoryUtils
