+++ platform-sdk/swirlds-merkledb/src/main/java/com/swirlds/merkledb/collections/HashListByteBuffer.java
@@ -12,7 +12,6 @@ import com.swirlds.base.utility.ToStringBuilder;
 import com.swirlds.config.api.Configuration;
 import com.swirlds.merkledb.config.MerkleDbConfig;
 import com.swirlds.merkledb.utilities.HashTools;
-import com.swirlds.merkledb.utilities.MemoryUtils;
 import com.swirlds.merkledb.utilities.MerkleDbFileUtils;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.io.IOException;
@@ -25,6 +24,7 @@ import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicLong;
 import org.hiero.base.crypto.Hash;
+import org.hiero.base.utility.MemoryUtils;
 
 /**
  * An implementation of {@link HashList} which makes use of an expanding, dynamic list of {@link ByteBuffer}s
